added SIMD (bugged)

+++ convolution.h
@@ -2,5 +2,5 @@
 #include <stdlib.h>
 #include <math.h>
 
-void calc_RGB_energy(int n, int m, int* channels, int* result);
-int* padd0_image(int n, int m, unsigned char* channels);
+void calc_RGB_energy(int n, int m, short* channels, int* result);
+short* padd0_image(int n, int m, unsigned char* channels);
